Rearranged project structure
