+++ main.py
@@ -1,0 +1,9 @@
+from stages.artefacts_stage import ArtefactsStage
+from stages.process_sequence import ProcessSequence
+
+if __name__ == '__main__':
+    ps = ProcessSequence(input_path='/home/quadro/videoproc/datasets/test_2_5K_frames/')
+    ps.add(ArtefactsStage())
+    ps.execute()
+
+    # assemble_video_lossless(imgs_path='/home/quadro/videoproc/video-processing/sr_stage_output/', framerate=25,filename='hockey_nlm_cb_sr')
