+++ run_scripts/run_artefacts.sh
@@ -5,9 +5,6 @@ while [ "$1" != "" ]; do
         -i | --input_path )     shift
                                 input_path="$1"
                                 ;;
-        -s | --sigma )          shift
-                                sigma="$1"
-                                ;;
         -o | --output_path )    shift
                                 output_path="$1"
                                 ;;
@@ -22,11 +19,6 @@ if [ -z "$input_path" ]; then
     exit 1
 fi
 
-if [ -z "$sigma" ]; then
-    echo "sigma unset"
-    exit 1
-fi
-
 if [ -z "$output_path" ]; then
     echo "output_path unset"
     exit 1
@@ -34,12 +26,11 @@ fi
 
 
 echo "input_path: $input_path";
-echo "sigma: $sigma";
 echo "output_path: $output_path";
 
+source ~/anaconda3/etc/profile.d/conda.sh
+conda activate artefacts
 
-conda activate fastdvdnet
-
-python3 ./src/nets/FastDVDNet/test_fastdvdnet.py --test_path $input_path --noise_sigma $sigma --save_path $output_path --model_file ./src/nets/FastDVDNet/model.pth
+python3 ./nets/infer_artefacts.py -i $input_path -o $output_path
 
 exit 0
