+++ stages/artefacts_stage.py
@@ -1,13 +1,11 @@
-import subprocess
-import os
 import shutil
+import subprocess
 from pathlib import Path
 
-from sequence_stage_base import SequenceStage
-import utils.logger as ps_logger
+import utils.logger as logger
+from stages.sequence_stage_base import SequenceStage
 
-
-logger = ps_logger.get_logger(__name__) 
+logger = logger.get_logger(__name__)
 
 
 class ArtefactsStage(SequenceStage):
@@ -20,8 +18,7 @@ class ArtefactsStage(SequenceStage):
         shutil.rmtree(self._output_path, ignore_errors=True)
         Path(self._output_path).mkdir(exist_ok=True)
 
-        base_path = Path(__file__).parent.absolute()
-        script_path = str(base_path / 'run_scripts' / 'run_artefacts.sh')
+        script_path = '../run_scripts/run_artefacts.sh'
         parameters = f'--input_path {self._input_path} --output_path {self._output_path}'
         run_cmd = script_path + ' ' + parameters
